Renamed EntityLink as SmartLink.
