+++ src/components/values/link/SmartLink.vue
